updated currPut.php
Updated currPut to first check if the code for the currency input is a valid one by trying to find it in the rest countries api, if it is the same function gets the data associated with it such as locations and full currency name. Also moved to using the reponse templates.